% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFS-package.R
\docType{package}
\name{BFS-package}
\alias{BFS}
\alias{BFS-package}
\title{BFS: Get Data from the Swiss Federal Statistical Office}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Search and download data from the Swiss Federal Statistical Office (BFS) APIs \url{https://www.bfs.admin.ch/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://felixluginbuhl.com/BFS/}
  \item \url{https://github.com/lgnbhl/BFS/}
  \item Report bugs at \url{https://github.com/lgnbhl/BFS/issues}
}

}
\author{
\strong{Maintainer}: Felix Luginbuhl \email{felix.luginbuhl@protonmail.ch} (\href{https://orcid.org/0009-0008-6625-2899}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Janosch Brenzel-Weiss [contributor]
  \item Joao Martins [contributor]
  \item Philipp Baumann \email{baumann-philipp@protonmail.com} (\href{https://orcid.org/0000-0002-3194-8975}{ORCID}) [contributor]
}

}
\keyword{internal}
