% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_download_geodata.R
\name{bfs_download_geodata}
\alias{bfs_download_geodata}
\title{Download a geographic file from the Swiss geo-portal}
\usage{
bfs_download_geodata(
  collection_id = collection_id,
  output_dir = tempdir(),
  overwrite = FALSE,
  create_json = FALSE,
  bbox = NULL,
  asset_names = NULL,
  datetime = NULL,
  feature_id = NULL
)
}
\arguments{
\item{collection_id}{collection_id}

\item{output_dir}{output_dir}

\item{overwrite}{overwrite}

\item{create_json}{create_json}

\item{bbox}{bbox}

\item{asset_names}{asset_names}

\item{datetime}{datetime}

\item{feature_id}{feature_id}
}
\description{
Download assets from the STAC API on the geodata catalog 
of the Swiss Confederation (\url{https://data.geo.admin.ch/api/stac/v0.9/}).
}
\details{
The use of the data and services is free but
subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
