% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_asset_metadata.R
\name{bfs_get_asset_metadata}
\alias{bfs_get_asset_metadata}
\title{Get asset metadata in a given language}
\usage{
bfs_get_asset_metadata(
  number_asset = NULL,
  number_bfs = NULL,
  language = c("de", "fr", "it", "en")
)
}
\arguments{
\item{number_asset}{The asset number of a dataset}

\item{number_bfs}{The BFS number of a dataset.}

\item{language}{character The language of a BFS catalog, i.e. "de", "fr", "it" or "en".}
}
\value{
list Returns a list containing asset metadata information. Returns NULL if no connection.
}
\description{
This function uses the DAM API \url{https://dam-api.bfs.admin.ch/hub/swagger-ui/index.html}
to get the metadata of a BFS file by asset number or BFS number in a given language.
}
\seealso{
[bfs_download_asset()]
}
