% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_catalog_data.R
\name{bfs_get_catalog_data}
\alias{bfs_get_catalog_data}
\title{Get the BFS data catalog}
\usage{
bfs_get_catalog_data(language = "de")
}
\arguments{
\item{language}{character The language of a BFS catalog.}
}
\value{
A data frame

A tbl_df (a type of data frame; see tibble or
dplyr packages).

\describe{
  \item{title}{A character column with the title of the BFS dataset}
  \item{language}{A character column with the language of the BFS dataset}
  \item{published}{A character column with a number for the text from each
  Wikisource page downloaded}
  \item{url_bfs}{A character column with the URL of the related BFS 
  webpage}
  \item{url_px}{A character column with the URL of the PX file}
}
}
\description{
This function scraps the RSS Feed of the Swiss Federal Statistical Office data catalog.
}
\examples{
\donttest{bfs_get_catalog_data(language = "de")}

}
\seealso{
\code{\link{bfs_get_data}}
}
