% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_catalog_geodata.R
\name{bfs_get_catalog_geodata}
\alias{bfs_get_catalog_geodata}
\title{Get the geodata catalog of the Swiss Confederation}
\usage{
bfs_get_catalog_geodata(include_metadata = TRUE)
}
\arguments{
\item{include_metadata}{boolean TRUE to loop on each geodata to add metadata.}
}
\description{
Display geo-information catalog of the Swiss Confederation (\url{https://data.geo.admin.ch/}),
including some geographic datasets provided by the Swiss Federal Statistical Office. 
Note that this geodata catalog is not hosted by the Swiss Federal Statistical Office.
}
\details{
For now only Stac API datasets are accessible.
}
