% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bfs_get_dataset}
\alias{bfs_get_dataset}
\title{Get BFS PC-Axis files as data frame}
\usage{
bfs_get_dataset(url_px)
}
\arguments{
\item{url_px}{The url link to download the PC-Axis file.}
}
\description{
The PC-Axis file dowloaded is saved in the working directory.
}
\examples{
\dontrun{df_en <- bfs_get_metadata(language = "en")}
\dontrun{bfs_search("education", df_en)}
\dontrun{bfs_get_dataset(df_en$url_px[3])}

}
