% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_dataset_comments.R
\name{bfs_get_dataset_comments}
\alias{bfs_get_dataset_comments}
\title{Get the comments/footnote of a BFS dataset in a given language}
\usage{
bfs_get_dataset_comments(
  url_bfs = NULL,
  language = "de",
  number_bfs = NULL,
  query = "all",
  clean_names = FALSE
)
}
\arguments{
\item{url_bfs}{The URL page of a dataset.}

\item{language}{Language of the dataset to be translated if exists.}

\item{number_bfs}{The BFS number of a dataset.}

\item{query}{a list object of the variables to query}

\item{clean_names}{Clean column names using \code{janitor::clean_names()}}
}
\description{
Get the comments/footnote of a BFS dataset in a given language
}
\seealso{
\code{\link{bfs_get_dataset}}
}
