% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bfs_get_dataset_meta}
\alias{bfs_get_dataset_meta}
\title{Get the metadata of a BFS data frame}
\usage{
bfs_get_dataset_meta(url_px, force = FALSE)
}
\arguments{
\item{url_px}{The url link to download the PC-Axis file.}

\item{force}{Force download to download data even if already downloaded today.

The data frame is saved in a temporary folder by default.}
}
\description{
Returns all the metadata in a data frame/tibble from a given BFS 
PC-Axis file. If the path of the cache argument is not provided, the 
downloaded BFS dataset will be saved in the default cache 
folder of the {pins} package.
}
\examples{
\donttest{meta_en <- bfs_get_metadata(language = "en")}
\donttest{bfs_meta_edu <- bfs_search(data = meta_en, string = "university students")}
\donttest{bfs_get_dataset_meta(bfs_meta_edu$url_px[1])}

}
