% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_sse_metadata.R
\name{bfs_get_sse_metadata}
\alias{bfs_get_sse_metadata}
\title{Get Metadata (Codelists) from Swiss Stats Explorer (SSE) API}
\usage{
bfs_get_sse_metadata(number_bfs, language = "de")
}
\arguments{
\item{number_bfs}{The BFS number (FSO number) of a dataset.}

\item{language}{Language of the codelists ("de", "fr", "it", "en").}
}
\value{
A tibble containing the codelists of the specified BFS dataset.
}
\description{
Retrieve metadata codelists for a specified BFS dataset from the Swiss Stats Explorer (SSE) API.
This function is required to built queries for \link{bfs_get_sse_data}.
}
\examples{
bfs_get_sse_metadata("DF_SSV_POL_LEG", language = "de")
}
