% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_sse_helpers.R
\name{bfs_get_sse_url}
\alias{bfs_get_sse_url}
\title{Get the URL for the Swiss Stats Explorer (SSE) API based on the BFS number}
\usage{
bfs_get_sse_url(number_bfs, metadata = FALSE)
}
\arguments{
\item{number_bfs}{The BFS number (FSO number) of a dataset.}

\item{metadata}{Logical; if TRUE, returns the URL for metadata instead of data.}
}
\value{
A character string containing the URL for the API of the specified BFS number.
}
\description{
Get the URL for the Swiss Stats Explorer (SSE) API based on the BFS number
}
