% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bfs_open_dir}
\alias{bfs_open_dir}
\title{Open Temp folder containing all downloaded BFS datasets}
\usage{
bfs_open_dir(bfs_data_dir = paste0(pins::board_local_storage()))
}
\arguments{
\item{bfs_data_dir}{A string}
}
\description{
Opens the folder which contains all the BFS datasets downloaded
using the bfs_get_dataset() function. As it is a temp folder, the
BFS datasets downloaded are cleared when the R session restarts.
}
\examples{
bfs_open_dir()

}
\seealso{
\code{\link{bfs_get_dataset}}
}
