% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_search.R
\name{bfs_search}
\alias{bfs_search}
\title{Search in BFS metadata  "list of the cubes"}
\usage{
bfs_search(word = NULL, word2 = NULL, langage = "de")
}
\description{
Search in BFS metadata  "list of the cubes"
}
