% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_catalog.R
\name{get_bfs_catalog}
\alias{get_bfs_catalog}
\title{Get the Swiss Federal Statistical Office data catalog in a given language}
\usage{
get_bfs_catalog(language = c("de", "fr", "it", "en"))
}
\arguments{
\item{language}{character The language of a BFS catalog.}
}
\value{
A data frame
}
\description{
This function scraps the RSS Feed of the Swiss Federal Statistical Office data catalog.
}
\examples{
\donttest{get_bfs_catalog(language = "de")}

}
\seealso{
\code{\link{bfs_get_metadata}}
}
