% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_bfs_metadata}
\alias{get_bfs_metadata}
\title{Get metadata of a give url page}
\usage{
get_bfs_metadata(url)
}
\arguments{
\item{url}{character The url of a BFS webpage}
}
\value{
A data frame
}
\description{
Get metadata of a give url page
}
