% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_bfs_metadata_all}
\alias{get_bfs_metadata_all}
\title{Looping on a list of BFS webpages urls}
\usage{
get_bfs_metadata_all(i)
}
\arguments{
\item{i}{character The url of a BFS webpage}
}
\value{
A data frame
}
\description{
Looping on a list of BFS webpages urls
}
