% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_sse_helpers.R
\name{get_xml_name}
\alias{get_xml_name}
\title{Get the name from an XML node for a specified language, with fallback}
\usage{
get_xml_name(node, language)
}
\arguments{
\item{node}{An XML node from which to extract the name.}

\item{language}{The desired language code ("de", "fr", "it", "en").}
}
\value{
A character string containing the name in the specified language or the first available name.
}
\description{
If the specified language is not found, it falls back to the first available name.
}
