% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{register_bzn}
\alias{register_bzn}
\title{Swiss official commune register BZN}
\format{
## `register_bzn`
A data frame with 143 rows and 3 columns:
\describe{
  \item{GDEKT}{Kantonskuerzel}
  \item{GDEBZNR}{Bezirksnummer}
  \item{GDEBZNA}{Bezirksname}
}
}
\source{
<https://www.bfs.admin.ch/bfs/fr/home/bases-statistiques/repertoire-officiel-communes-suisse.html>
}
\usage{
register_bzn
}
\description{
The official commune register is structured according to cantons and
districts or comparable administrative entities. Various federal, cantonal
and communal governments and private businesses use this register when
identifying and referring to communes.
}
\details{
The Federal Statistical Office assigns a number to each commune and creates,
administers and publishes the Swiss official commune register.
}
\keyword{datasets}
