% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{register_dic}
\alias{register_dic}
\title{Swiss official commune register Dictionary}
\format{
## `register_dic`
A data frame with 27 rows and 5 columns:
\describe{
  \item{language}{Language}
  \item{abbreviation}{Abbreviation}
  \item{title}{Title}
  \item{spec}{Specifications}
  \item{notes}{General notes}
}
}
\source{
<https://www.bfs.admin.ch/bfs/fr/home/bases-statistiques/repertoire-officiel-communes-suisse.html>
}
\usage{
register_dic
}
\description{
The official commune register is structured according to cantons and
districts or comparable administrative entities. Various federal, cantonal
and communal governments and private businesses use this register when
identifying and referring to communes.
}
\details{
The Federal Statistical Office assigns a number to each commune and creates,
administers and publishes the Swiss official commune register.
}
\keyword{datasets}
