% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{register_gde}
\alias{register_gde}
\title{Swiss official commune register GDE}
\format{
## `register_gde`
A data frame with 2,136 rows and 8 columns:
\describe{
  \item{GDEKT}{Kantonskuerzel}
  \item{GDEBZNR}{Bezirksnummer}
  \item{GDENR}{BFS-Gemeindenummer}
  \item{GDENAME}{Amtlicher Gemeindename}
  \item{GDENAMK}{Gemeindename, kurz}
  \item{GDEBZNA}{Bezirksname}
  \item{GDEKTNA}{Kantonsname}
  \item{GDEMUTDAT}{Datum der letzten Anderung}
}
}
\source{
<https://www.bfs.admin.ch/bfs/fr/home/bases-statistiques/repertoire-officiel-communes-suisse.html>
}
\usage{
register_gde
}
\description{
The official commune register is structured according to cantons and
districts or comparable administrative entities. Various federal, cantonal
and communal governments and private businesses use this register when
identifying and referring to communes.
}
\details{
The Federal Statistical Office assigns a number to each commune and creates,
administers and publishes the Swiss official commune register.
}
\keyword{datasets}
