% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{register_kt}
\alias{register_kt}
\title{Swiss official commune register KT}
\format{
## `register_kt`
A data frame with 26 rows and 3 columns:
\describe{
  \item{KTNR}{Kantonsnummer}
  \item{GDEKT}{Kantonskuerzel}
  \item{GDEKTNA}{Kantonsname}
}
}
\source{
<https://www.bfs.admin.ch/bfs/fr/home/bases-statistiques/repertoire-officiel-communes-suisse.html>
}
\usage{
register_kt
}
\description{
The official commune register is structured according to cantons and 
districts or comparable administrative entities. Various federal, cantonal 
and communal governments and private businesses use this register when 
identifying and referring to communes.
}
\details{
The Federal Statistical Office assigns a number to each commune and creates, 
administers and publishes the Swiss official commune register.
}
\keyword{datasets}
