% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{register_kt_seeanteile}
\alias{register_kt_seeanteile}
\title{Swiss official commune register KT Seeanteile}
\format{
## `register_kt_seeanteile`
A data frame with 9 rows and 5 columns:
\describe{
  \item{GDENR}{BFS-Gemeindenummer}
  \item{GDENAME}{Amtlicher Gemeindename}
  \item{KTNR}{Kantonsnummer}
}
}
\source{
<https://www.bfs.admin.ch/bfs/fr/home/bases-statistiques/repertoire-officiel-communes-suisse.html>
}
\usage{
register_kt_seeanteile
}
\description{
The official commune register is structured according to cantons and 
districts or comparable administrative entities. Various federal, cantonal 
and communal governments and private businesses use this register when 
identifying and referring to communes.
}
\details{
The Federal Statistical Office assigns a number to each commune and creates, 
administers and publishes the Swiss official commune register.
}
\keyword{datasets}
