% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_catalog.R
\name{bfs_get_catalog}
\alias{bfs_get_catalog}
\title{Get the BFS data or table catalog}
\usage{
bfs_get_catalog(language = "de", type = "data")
}
\arguments{
\item{language}{character The language of a BFS catalog.}

\item{type}{character A BFS catalog}
}
\value{
A data frame. Returns NULL if no connection.
}
\description{
This function scraps a given RSS Feed of the Swiss Federal Statistical Office.
}
\seealso{
\code{\link{bfs_get_data}}
}
