% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_catalog_data.R
\name{bfs_get_catalog_data}
\alias{bfs_get_catalog_data}
\title{Get the BFS data catalog}
\usage{
bfs_get_catalog_data(language = c("de", "fr", "it", "en"))
}
\arguments{
\item{language}{character The language of a BFS catalog.}
}
\value{
A data frame
}
\description{
This function scraps the RSS Feed of the Swiss Federal Statistical Office data catalog.
}
\examples{
\donttest{bfs_get_catalog_data(language = "de")}

}
\seealso{
\code{\link{bfs_get_data}}
}
