% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_catalog_tables.R
\name{bfs_get_catalog_tables}
\alias{bfs_get_catalog_tables}
\title{Get the BFS tables catalog}
\usage{
bfs_get_catalog_tables(language = "de")
}
\arguments{
\item{language}{character The language of a BFS catalog.}
}
\value{
A data frame
}
\description{
This function scraps the RSS Feed of the Swiss Federal Statistical Office tables catalog.
The tables are mostly datasets that are accessible as Excel files.
}
\examples{
\donttest{bfs_get_catalog_tables(language = "de")}

}
