% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_metadata.R
\name{bfs_get_metadata}
\alias{bfs_get_metadata}
\title{Download BFS metadata in a given language}
\usage{
bfs_get_metadata(language = "de")
}
\arguments{
\item{language}{character The language of the metadata.}
}
\value{
A tibble
}
\description{
Returns a tibble containing the titles, publication dates,
observation periods, data source, metadata webpage urls and download link urls
in a given language of the current public BFS datasets available.
}
\seealso{
\code{\link{bfs_get_dataset}}
}
