% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bfs_get_metadata}
\alias{bfs_get_metadata}
\title{Get all BFS metadata in a given language}
\usage{
bfs_get_metadata(language = "de")
}
\arguments{
\item{language}{character The language of the metadata

Languages availables are German ("de", as default), French ("fr"),
Italian ("it") and English ("en"). Note that Italian and English BFS
metadata doesn't give access to all the BFS datasets availables online.}
}
\value{
A tibble
}
\description{
Returns a tibble containing the titles, publication dates,
observation periods, metadata urls and download urls of
available BFS datasets in a given language.
}
\examples{
\donttest{df_en <- bfs_get_metadata(language = "en")}

}
