% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_metadata.R
\name{bfs_get_metadata}
\alias{bfs_get_metadata}
\title{Get metadata of a BFS data in a given language}
\usage{
bfs_get_metadata(number_bfs = NULL, url_bfs = NULL, language = "de")
}
\arguments{
\item{number_bfs}{The BFS number of a dataset.}

\item{url_bfs}{The URL page of a dataset.}

\item{language}{Language of the dataset to be translated if exists.}
}
\value{
A tbl_df (a type of data frame; see tibble or dplyr packages).
}
\description{
Get the metadata of a BFS dataset using the PXWEB API v1.
You should choose either the bfs number of the bfs offical url of a given dataset.
}
