% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs_get_metadata.R
\name{bfs_get_metadata}
\alias{bfs_get_metadata}
\title{Get metadata of a BFS data in a given language}
\usage{
bfs_get_metadata(url_bfs = NULL, language = "de", number_bfs = NULL)
}
\arguments{
\item{url_bfs}{The URL page of a dataset.}

\item{language}{Language of the dataset to be translated if exists.}

\item{number_bfs}{The BFS number of a dataset.}
}
\value{
A tbl_df (a type of data frame; see tibble or dplyr packages).
}
\description{
Get the metadata of a BFS dataset using the BFS offical API v1. 
You should choose either the bfs number of the bfs offical url of a given dataset.
}
\examples{
\donttest{bfs_get_metadata(number_bfs = "px-x-2105000000_501", language = "fr")}

}
