% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bfs_open_dir}
\alias{bfs_open_dir}
\title{Open folder containing all downloaded BFS datasets}
\usage{
bfs_open_dir(path = pins::board_local_storage())
}
\arguments{
\item{path}{The local folder to use as a cache, default to {pins} cache.}
}
\description{
Opens the folder which contains all the BFS datasets downloaded
relatively to their path argument, using the {pins} package. If
no path argument is provided, the downloaded BFS datasets will be
saved in the default cache folder of the {pins} package.
}
\examples{
\donttest{bfs_open_dir()}

}
\seealso{
\code{\link{bfs_get_dataset}}
}
